requirements.txt
litellm
python-dotenv
networkx
matplotlib
sentence-transformers
scikit-learn
numpy
